requirements.txt
alembic==1.14.1
attrs==24.2.0
Automat==24.8.1
blinker==1.8.2
certifi==2021.5.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==43.0.3
cssselect==1.2.0
CTkTable==1.1
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
defusedxml==0.7.1
dnspython==2.6.1
docker==7.0.0
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
greenlet==3.1.1
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
lxml==5.3.0
Mako==1.3.9
MarkupSafe==2.1.5
matplotlib==3.8.4
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
parsel==1.9.1
pillow==10.2.0
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pygame==2.5.2
pymongo==4.6.2
pyOpenSSL==24.2.1
pyparsing==3.1.2
pypyodbc==1.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
queuelib==1.7.0
requests==2.31.0
requests-file==2.1.0
scapy==2.5.0
Scrapy==2.11.2
service-identity==24.1.0
setuptools==69.1.0
six==1.16.0
SQLAlchemy==2.0.37
tldextract==5.1.2
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==1.26.6
vboxapi==1.0
w3lib==2.2.1
Werkzeug==3.0.3
wheel==0.42.0
zope.interface==7.1.0
